pyproject.toml
[project]
name = "insight-deck"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dynaconf>=3.2.11",
    "fastapi[standard]>=0.116.1",
    "feedparser>=6.0.11",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.1.8",
    "langchain[google-genai]>=0.3.26",
    "python-dotenv>=1.1.1",
]
